hf_deploy/requirements.txt
fastapi
uvicorn
yt-dlp>=2024.08.06
requests
pydantic
certifi
dnspython
